+++ operator/api/v1beta1/zz_generated.deepcopy.go
@@ -988,8 +988,13 @@ func (in *DaemonSpec) DeepCopyInto(out *DaemonSpec) {
 		}
 	}
 	in.Resources.DeepCopyInto(&out.Resources)
-	if in.ContainerSecurity != nil {
-		in, out := &in.ContainerSecurity, &out.ContainerSecurity
+	if in.Capabilities != nil {
+		in, out := &in.Capabilities, &out.Capabilities
+		*out = new(v1.SecurityContext)
+		(*in).DeepCopyInto(*out)
+	}
+	if in.AISContainerSecurityContext != nil {
+		in, out := &in.AISContainerSecurityContext, &out.AISContainerSecurityContext
 		*out = new(v1.SecurityContext)
 		(*in).DeepCopyInto(*out)
 	}