+++ operator/pkg/resources/cmn/utils.go
@@ -93,10 +93,3 @@ func IterPtr[Slice ~[]E, E any](s Slice) iter.Seq[*E] {
 		}
 	}
 }
-
-func ValueOrDefault[T any](value, defaultValue *T) *T {
-	if value == nil {
-		return defaultValue
-	}
-	return value
-}