+++ operator/pkg/resources/proxy/statefulset.go
@@ -107,7 +107,7 @@ func proxyPodSpec(ais *aisv1.AIStore) *corev1.PodSpec {
 				Env:             NewAISContainerEnv(ais),
 				Ports:           cmn.NewDaemonPorts(&ais.Spec.ProxySpec),
 				Resources:       *cmn.NewResourceReq(ais, &ais.Spec.ProxySpec.Resources),
-				SecurityContext: ais.Spec.ProxySpec.ContainerSecurity,
+				SecurityContext: cmn.GetAISSecurityContext(&ais.Spec.ProxySpec),
 				VolumeMounts:    newVolumeMounts(ais),
 				StartupProbe:    cmn.NewStartupProbe(ais, aisapc.Proxy),
 				LivenessProbe:   cmn.NewLivenessProbe(ais, aisapc.Proxy),
@@ -117,13 +117,9 @@ func proxyPodSpec(ais *aisv1.AIStore) *corev1.PodSpec {
 		Affinity:           cmn.CreateAISAffinity(ais.Spec.ProxySpec.Affinity, SelectorLabels(ais)),
 		NodeSelector:       ais.Spec.ProxySpec.NodeSelector,
 		ServiceAccountName: cmn.ServiceAccountName(ais),
-		// By default, Kubernetes sets non-nil `SecurityContext`. So we have to do that too,
-		// otherwise during comparison we will always fail (nil vs non-nil).
-		//
-		// See: https://github.com/kubernetes/kubernetes/blob/fa03b93d25a5a22d4f91e4c44f66fc69a6f69a35/pkg/apis/core/v1/defaults.go#L215-L236
-		SecurityContext: cmn.ValueOrDefault(ais.Spec.ProxySpec.SecurityContext, &corev1.PodSecurityContext{}),
-		Volumes:         newVolumes(ais),
-		Tolerations:     ais.Spec.ProxySpec.Tolerations,
+		SecurityContext:    cmn.GetPodSecurityContext(&ais.Spec.ProxySpec),
+		Volumes:            newVolumes(ais),
+		Tolerations:        ais.Spec.ProxySpec.Tolerations,
 	}
 	// Apply priority class if specified to prevent eviction during node pressure
 	if ais.Spec.PriorityClassName != nil {